processing.py created and functions moved there
